+++ tests/test_stac.py
@@ -6,7 +6,6 @@ from unittest.mock import patch
 import pytest
 from pystac.utils import str_to_datetime
 
-from stactools.hotosm.constants import COLLECTION_ID
 from stactools.hotosm.exceptions import AssetNotFoundError
 from stactools.hotosm.oam_metadata import OamMetadata
 from stactools.hotosm.stac import create_collection, create_item
@@ -22,8 +21,6 @@ def test_create_item(example_oam_image: OamMetadata):
     """Test Item creation."""
     item = create_item(example_oam_image)
     item.validate()
-
-    assert item.collection_id == COLLECTION_ID
 
     for datetime_prop in {"start_datetime", "end_datetime"}:
         assert isinstance(str_to_datetime(item.properties[datetime_prop]), dt.datetime)